+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "botzilla",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Discord bot for Bobzilla4tv server",
   "main": "bot.js",
   "scripts": {
